requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
celery==5.3.6
redis==5.0.1
playwright==1.42.0
beautifulsoup4==4.12.3
groq==0.5.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.0.2
